requirements.txt
asgiref==3.9.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dtaidistance==2.3.13
fonttools==4.60.0
kiwisolver==1.4.9
matplotlib==3.10.6
mysqlclient==2.2.7
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
rest-framework-simplejwt==0.0.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
